Add tiered storage so the database stops growing without bound
At 50 machines x 1 reading/second the database gains ~1.5-2 GB/day and never
gives any of it back. Two things break before that becomes a capacity question:
SQL Server Express caps a database at a hard 10 GB (reached in under a week),
and GET /api/v1/machines was an unbounded GROUP BY over the whole Telemetries
table that the dashboard polls every 5 seconds.

Replace the ever-growing raw tables with retention tiers plus pre-aggregation.
The lever for time-series data is reducing resolution, not relocating rows: a
hot/cold pair of identically shaped tables stores exactly as many bytes and
leaves the expensive dashboard queries scanning the hot side.

Storage tiers:
- Hot: Telemetries (72h) and SensorReadings (24h), full fidelity. Sensor
  readings are mechanically derived from telemetry, so a shorter window there
  drops two-thirds of all rows written at no loss of information.
- Warm/cold: per-minute buckets (30d) re-aggregated into per-hour buckets (2y).
  Buckets store sums and counts, never averages -- averaging per-minute
  averages is only correct when every bucket holds the same sample count, which
  stops being true the moment a machine drops a reading.
- Roster: MachineSummaries, one running row per machine, folded forward as each
  minute bucket closes. Turns the fleet query from "read every row ever
  ingested" into a 50-row read that costs the same on day 1000 as on day 1.

Physical layout: re-key both hot tables on (Timestamp, Id). A random GUID as
the clustered key scatters every insert to a random page, so ingestion degrades
continuously as the table grows; leading with time makes it an append, and
makes aggregation and retention contiguous range operations. The per-machine
"latest N" path moves to a covering index. Both get page compression.

DataLifecycleWorker aggregates then purges, with each bucket's aggregate, roster
fold and checkpoint advance in one transaction -- which is what makes the roster
safe to accumulate rather than recompute. Purge cutoffs are clamped to how far
aggregation actually reached, so a stalled rollup grows the database (recoverable)
instead of deleting rows nothing summarised (not).

Read queries select a tier by window width: up to 3 hours from raw rows and
exact to the second, wider windows from buckets. The dashboard's 6h and 24h
views drop from scanning millions of rows to tens of thousands.

Also bound the ingestion channel. It was unbounded, so a slow database grew the
heap until the process was OOM-killed along with everything it held; waiting on
a full channel instead delays the ack, fills the prefetch window, and leaves the
backlog in RabbitMQ where it is durable.

Steady-state size lands around 2 GB and stays there. Design rationale, capacity
arithmetic and the scale-out path (partition switching, columnstore, TSDB) are
in docs/DATA-LIFECYCLE.md.

+++ src/FactoryIoT.Domain/Analytics/MachineTelemetrySummary.cs
@@ -5,14 +5,16 @@ namespace FactoryIoT.Domain.Analytics;
 /// used to answer the "give me the whole fleet at a glance" question without pulling raw rows.
 /// </summary>
 /// <remarks>
-/// This is a read-model (an aggregate projection), not a persisted entity: it is produced by a
-/// <c>GROUP BY MachineId</c> over the <c>Telemetries</c> table, so it carries only aggregate
-/// values (count, extents, averages) rather than any single row. The shape is intentionally flat
-/// so it maps directly to a translatable SQL projection (see <c>TelemetryRepository</c>).
+/// This is a read-model (an aggregate projection), not a persisted entity. It is served from the
+/// <c>MachineSummaries</c> tier — one running row per machine, folded forward by the rollup job —
+/// so the query stays a fifty-row read instead of the unbounded <c>GROUP BY MachineId</c> over the
+/// whole raw table it used to be. The figures are therefore lifetime totals that survive
+/// retention: <c>MaxTemperature</c> is the hottest the machine has ever run, not the hottest
+/// within whatever raw history is still on disk.
 /// </remarks>
 public sealed record MachineTelemetrySummary(
     string MachineId,
-    int SampleCount,
+    long SampleCount,
     DateTimeOffset FirstSeen,
     DateTimeOffset LastSeen,
     double MinTemperature,
